Bugfix: Use default issue_type strings
It looks like this has possibly changed in the API over time? the
existing values are rejected by the server as being invalid.

I've simply changed them to "valid" locator values from an almost vanilla
docker-compose instance.  But I'm not sure if this should really be using
statically included strings at all, but rather using ones specified in
the report_portal.yml or scenario tags instead?

I'll leave this one for the project maintainers to ponder